api cleanup, sse edition

+++ ref/blake2.h
@@ -52,7 +52,7 @@ extern "C" {
     uint32_t t[2];
     uint32_t f[2];
     uint8_t  buf[BLAKE2S_BLOCKBYTES];
-    uint32_t buflen;
+    size_t   buflen;
     uint8_t  last_node;
   } blake2s_state;
 
@@ -62,7 +62,7 @@ extern "C" {
     uint64_t t[2];
     uint64_t f[2];
     uint8_t  buf[BLAKE2B_BLOCKBYTES];
-    uint32_t buflen;
+    size_t   buflen;
     uint8_t  last_node;
   } blake2b_state;
 
@@ -71,7 +71,7 @@ extern "C" {
     blake2s_state S[8][1];
     blake2s_state R[1];
     uint8_t       buf[8 * BLAKE2S_BLOCKBYTES];
-    uint32_t      buflen;
+    size_t        buflen;
   } blake2sp_state;
 
   typedef struct blake2bp_state__
@@ -79,7 +79,7 @@ extern "C" {
     blake2b_state S[4][1];
     blake2b_state R[1];
     uint8_t       buf[4 * BLAKE2B_BLOCKBYTES];
-    uint32_t      buflen;
+    size_t        buflen;
   } blake2bp_state;
